Get read deadline from connection
This changes the internal calls to
mock.Connection readers such that they
use the current readDeadline set
in the Connection. Earlier it got the value
from what was passed in the read call - which would be an old value if the value
is changed while one read is blocked.

It is left as a TODO to have a different
read deadline for remote reads.