Refactor to use external packages tada/catch and tada/jsonstream
tada/catch is a separate module that sprung from the ideas of
mqtt-nats/pio. It has since undergone several improvements. The same
story applies to the tada/jsonstream module which originated as a
package in mqtt-nats and matured into a module of its own.

This commit removes the pio and jsonstream packages in favor of using
the new external modules.