Use workers with throttle for incoming connections
Previously, the number of go routines needed to accommodate incoming
connections was dictated by the `Accept()` call. Each new connection
would start a new go routine.

This commit will instead push all connections on an `incoming` queue and
then have a number of workers serve that queue. The number of workers is
automatically throttled so that new workers are started when the number
of available workers is too low and workers die off when that number is
too high.